Solucionado error cuando se cargan tablas vacias

+++ Python/TLS_Script.py
@@ -25,57 +25,59 @@ if (len(b)==6) and (int(b[4:])<13) and (int(b[4:])>0) and (int(b[:4]) <= int(yea
 			else:
 				return x
 
-		datos.columns = list(map(lambda x: x.lower().replace(" ", "_").replace("-","_").replace("+","plus").replace("(","").replace(")","")
-		.replace("á","a").replace("é","e").replace("í","i").replace("ó","o").replace("ú","u").replace("mes","month"), datos.columns))
-		
-		datos.horas_estabilizacion = list(map(lambda x: cambiarNAN(x), datos.horas_estabilizacion))
-		datos['auditoria']=pd.Series([datetime.now() for x in range(len(datos.index))])
-		
-		datos1 = datos
-		
-		datos = datos.drop_duplicates(subset = ["month", "project", "id_employee"], keep = 'first')
-		
-		datos = datos.dropna(subset = ["month", "project", "id_employee"])
-		m = datos.merge(datos1, how = "outer", suffixes = ['','_'], indicator = True)
-		mer = m.loc[m._merge.eq('right_only')]
-		mer = mer.drop("_merge", axis = 1)
-		
-		if  len(mer) > 0: 
-			print("Existen registros rechazados, compruebe el xlsx generado con los rechazados")
+		if len(datos) > 0:
+			datos.columns = list(map(lambda x: x.lower().replace(" ", "_").replace("-","_").replace("+","plus").replace("(","").replace(")","")
+			.replace("á","a").replace("é","e").replace("í","i").replace("ó","o").replace("ú","u").replace("mes","month"), datos.columns))
 			
-		datos = datos.reset_index(drop = True)
-		
-		config = configparser.ConfigParser()
-		config.read("configuracion.ini")
-		
-		usuario = sys.argv[1]
-
-		usuario = usuario.upper()
-		password = config[usuario]["password"]
-		user = config[usuario]["user"]
-		host = config[usuario]["host"]
-		dataBase = config[usuario]["dataBase"]
-
-		engine = sqlalchemy.create_engine('mysql+pymysql://'+user+':'+password+'@'+host+'/'+dataBase)
-		
-		engine.execute("delete from black_margin.tls where month = "+b+";")
-		try:
-			datos.to_sql("tls", engine, if_exists = "append", index = False)
-		except:
-			print("Error en el formato de la tabla, revise el excel y vuelva a realizar la tabla.")
-
-		if usuario=='SERVIDOR':
-			writer = pd.ExcelWriter('C:/Users/MicroStrategyBI/Desktop/black_margin_backup/historicos_black_margin/tls_acumulado.xlsx', engine='xlsxwriter')
-			datos.to_excel(writer, index=False)
-			writer.save()
-			if path.exists('C:/Users/MicroStrategyBI/Desktop/black_margin_backup/rechazados_black_margin/rechazados_tls.xlsx'):
-				rechazados = pd.read_excel('C:/Users/MicroStrategyBI/Desktop/black_margin_backup/rechazados_black_margin/rechazados_tls.xlsx')
-				mer = pd.concat([rechazados, mer])
+			datos.horas_estabilizacion = list(map(lambda x: cambiarNAN(x), datos.horas_estabilizacion))
+			datos['auditoria']=pd.Series([datetime.now() for x in range(len(datos.index))])
+			
+			datos1 = datos
+			
+			datos = datos.drop_duplicates(subset = ["month", "project", "id_employee"], keep = 'first')
+			
+			datos = datos.dropna(subset = ["month", "project", "id_employee"])
+			m = datos.merge(datos1, how = "outer", suffixes = ['','_'], indicator = True)
+			mer = m.loc[m._merge.eq('right_only')]
+			mer = mer.drop("_merge", axis = 1)
+			
+			if  len(mer) > 0: 
+				print("Existen registros rechazados, compruebe el xlsx generado con los rechazados")
 				
-			writer = pd.ExcelWriter('C:/Users/MicroStrategyBI/Desktop/black_margin_backup/rechazados_black_margin/rechazados_tls.xlsx', engine='xlsxwriter')
-			mer.to_excel(writer, index=False)
-			writer.save()
+			datos = datos.reset_index(drop = True)
 			
+			config = configparser.ConfigParser()
+			config.read("configuracion.ini")
+			
+			usuario = sys.argv[1]
+
+			usuario = usuario.upper()
+			password = config[usuario]["password"]
+			user = config[usuario]["user"]
+			host = config[usuario]["host"]
+			dataBase = config[usuario]["dataBase"]
+
+			engine = sqlalchemy.create_engine('mysql+pymysql://'+user+':'+password+'@'+host+'/'+dataBase)
+			
+			engine.execute("delete from black_margin.tls where month = "+b+";")
+			try:
+				datos.to_sql("tls", engine, if_exists = "append", index = False)
+			except:
+				print("Error en el formato de la tabla, revise el excel y vuelva a realizar la tabla.")
+
+			if usuario=='SERVIDOR':
+				writer = pd.ExcelWriter('C:/Users/MicroStrategyBI/Desktop/black_margin_backup/historicos_black_margin/tls_acumulado.xlsx', engine='xlsxwriter')
+				datos.to_excel(writer, index=False)
+				writer.save()
+				if path.exists('C:/Users/MicroStrategyBI/Desktop/black_margin_backup/rechazados_black_margin/rechazados_tls.xlsx'):
+					rechazados = pd.read_excel('C:/Users/MicroStrategyBI/Desktop/black_margin_backup/rechazados_black_margin/rechazados_tls.xlsx')
+					mer = pd.concat([rechazados, mer])
+					
+				writer = pd.ExcelWriter('C:/Users/MicroStrategyBI/Desktop/black_margin_backup/rechazados_black_margin/rechazados_tls.xlsx', engine='xlsxwriter')
+				mer.to_excel(writer, index=False)
+				writer.save()
+		else:
+			print("El excel no contiene datos.")
 	else:
 		print("El archivo que intenta consultar no existe porque la fecha no coincide o no existe en este directorio")	
 else:
